Split test cases for cluster and standalone
